moved status list to constant class
